[Feat] : slider module 분리
